NEW/shipyard/requirements.txt
Django==1.5.1
Fabric==1.7.0
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
South==0.8.1
apache-libcloud==0.13.1
argparse==1.2.1
bpython==0.12
cmdln==1.1.2
django-crispy-forms==1.3.0
django-redis==3.3
django-rq==0.5.1
-e git+http://github.com/dotcloud/docker-py.git@ab0f98915427d34ebc7472a35fa7657cbf8164bc#egg=docker_py-master
maestro==0.1
paramiko==1.11.0
pycrypto==2.6
python-dateutil==2.1
pytz==2013b
redis==2.7.6
requests==1.2.3
rq==0.3.7
six==1.3.0
times==0.6.1
wsgiref==0.1.2
